fix: 로그인 반환값 수정

+++ buddyFriends/src/main/java/com/buddyFriends/buddyFriends/base/dto/PetDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class PetDto {
 
     @Nullable
-    private Long petId;
+    private Integer petId;
 
     @Nullable
     private String userId;
